requirements.txt
django>2.0
django-crispy-forms>=1.0.0
django-simple-menu